% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_covariates.R
\name{baseline_covariates}
\alias{baseline_covariates}
\title{{Baseline covariates: simulazione delle variabili confondenti}}
\usage{
baseline_covariates(
  n,
  mus = c(rep(0, 7L), 55, 26),
  sds = c(rep(1, 7L), 11, 4),
  corr_matrix
)
}
\arguments{
\item{{n}}{{Numero di casi che si vuole simulare}}

\item{{mus}}{{Vettore contenente le medie per ciascuna variabile che si vuole simulare}}

\item{{sds}}{{Vettore contenente le deviazioni standard}}

\item{{corr_matrix}}{{Matrice di correlazione tra le variabili che si vuole stimare}}
}
\value{
{Matrice il cui numero di righe è pari al numero di osservazioni che si vuole stimare e il numero di colonne è pari al numero di variabili che si vuole simulare}
}
\description{
Simulazione delle covariate a partire da una distribuzione normale multivariata. In particolare, verranno ad essere simulate delle variabili che dovrebbero simulare le nove covariate confondenti del caso studio: 7 variabili binarie (con una prevalenza identica a quelle del caso studio), e 2 variabili continue (con media e sd pari alle corrispettive del caso studio)
}
