% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisection.R
\name{bisection}
\alias{bisection}
\title{{Bisection: find the intercept for the propensity score model}}
\usage{
bisection(
  treatment_prevalence,
  lower = -500,
  upper = 500,
  maxError = 1e-04,
  maxIter = 10000,
  baseline_covariates,
  alpha
)
}
\arguments{
\item{{treatment_prevalence}}{{Desired treatment prevalence}}

\item{{lower}}{{Minimum value that the intercept can assume}}

\item{{upper}}{{Maximum value that the intercept can assume}}

\item{{maxIter}}{{Max number of iterations}}

\item{{maxError}}{{Max error}}

\item{{baseline_covariates}}{{Simulated baseline covariates, obtained through `baseline_covariates`}}

\item{{alpha}}{{Vector of alphas, necessary for using `define_p_treat`}}
}
\description{
To find the right intercept of the propensity score model to obtain a specific treatment prevalence
}
