% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix_statin.R
\name{corr_matrix_statin}
\alias{corr_matrix_statin}
\title{{corr_matrix_statin: estrazione della matrice di correlazione dal caso studio}}
\usage{
corr_matrix_statin(data)
}
\arguments{
\item{{data}}{{Dataset del caso studio sull'utilizzo di statine}}
}
\description{
{corr_matrix_statin: estrazione della matrice di correlazione dal caso studio}
}
