% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_alpha.R
\name{define_alpha}
\alias{define_alpha}
\title{{Define alpha: definire i coefficienti di regressione logistica per le variabili confondenti}}
\usage{
define_alpha(data)
}
\arguments{
\item{{data}}{{Base di dati iniziali, che sarebbe quella del mio caso studio}}
}
\description{
{Define alpha: definire i coefficienti di regressione logistica per le variabili confondenti}
}
