% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_beta.R
\name{define_beta}
\alias{define_beta}
\title{{Define beta: definire i coefficienti di regressione per la simulazione della variabile di outcome}}
\usage{
define_beta(data)
}
\arguments{
\item{{data}}{{Base di dati, ovvero quella del nostro caso studio}}
}
\description{
{Define beta: definire i coefficienti di regressione per la simulazione della variabile di outcome}
}
