% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_p_treat.R
\name{define_p_treat}
\alias{define_p_treat}
\title{{Simulate the propensity score for the simulated observations}}
\usage{
define_p_treat(int, alpha, baseline_covariates)
}
\arguments{
\item{{int}}{{Logistic regression model intercept, obtained through `bisection`}}

\item{{alpha}}{{Logistic regression model coefficients: must have the same length as the number of columns of the matrix obtained through `baseline_covariates`}}

\item{{baseline_covariates}}{{Simulated baseline covariates obtained throught `baseline_covariates`}}
}
\description{
Define the probability of receiving the treatment giving a series of simulated baseline covariates
}
