% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{{Simulation: simulazione}}
\usage{
simulation(
  n,
  n_datasets = 1000,
  seed = 42,
  treatment_prevalence,
  treatment_effect = -0.2,
  approach = c("crude", "IPW", "OW"),
  trim = NULL
)
}
\arguments{
\item{{n}}{{Numero di osservazioni che vogliamo simulare}}

\item{{n_datasets}}{{Numero di dataset che vogliamo simulare, default 1000}}

\item{{seed}}{{Seed per la randomizzazione, default 42}}

\item{{treatment_prevalence}}{{La treatment prevalence desiderata}}

\item{{treatment_effect}}{{Il treatment effect stabilito, default -0.2}}

\item{{approach}}{{Tipo di approccio che vogliamo utilizzare, i.e. crude, IPW, OW}}

\item{{trim}}{{Se si è scelto IPW, specificare il cutoff per l'eventuale trimming (1 indica no trimming)}}
}
\value{
{Una lista contenente le seguenti informazioni: livello di sbilanciamento delle covariate, il bias, il RMSE}
}
\description{
Simulare una determinata configurazione che vogliamo testare
}
