% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_results.R
\name{simulation_results}
\alias{simulation_results}
\title{{Simulation results}}
\usage{
simulation_results(sim_results, treatment_effect = -0.2)
}
\arguments{
\item{{sim_results}}{{Output della funzione `simulation`}}

\item{{treatment_effect}}{{Treatment effect stabilito, deve essere uguale a quello di simulation, default -0.2}}
}
\description{
Estrarre i risultati delle simulazioni in termini di bias, SMD e RMSE
}
