% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_covariates.R
\name{baseline_covariates}
\alias{baseline_covariates}
\title{{Baseline covariates: baseline covariates simulation}}
\usage{
baseline_covariates(
  n,
  mus = c(rep(0, 7L), 55, 26),
  sds = c(rep(1, 7L), 11, 4),
  corr_matrix
)
}
\arguments{
\item{{n}}{{Number of simulated cases}}

\item{{mus}}{{Vector of the means for each simulated variable}}

\item{{sds}}{{Vector of the sds for each simulated variable}}

\item{{corr_matrix}}{{Matrix of correlations between simulated variables}}
}
\value{
{A matrix in which each row corresponds to a simulated case and each column corresponds to a simulated variable}
}
\description{
Baseline covariates simulation, starting from a multivariate normal distribution. Particularly, the simulated variables would resemble the case study ones: 7 binary variables (with the same prevalence as in the case study), and 2 continuous variables (with the same mean and sd as in the case study)
}
