% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisection.R
\name{bisection}
\alias{bisection}
\title{{Bisection: definizione dell'intercetta per il modello di stima dei propensity score}}
\usage{
bisection(
  treatment_prevalence,
  lower = -500,
  upper = 500,
  maxError = 1e-04,
  maxIter = 10000,
  baseline_covariates,
  alpha
)
}
\arguments{
\item{{treatment_prevalence}}{{La treatment prevalence desiderata (deve andare da 0 a 1)}}

\item{{lower}}{{Valore minimo che l'intercetta può assumere}}

\item{{upper}}{{Valore massimo che l'intercetta può assumere}}

\item{{maxIter}}{{Massimo numero di iterazioni per la ricerca dell'intercetta}}

\item{{maxError}}{{Errore massimo}}

\item{{baseline_covariates}}{{Variabili confondenti simulate attraverso la funzione `baseline_covariates`}}

\item{{alpha}}{{Vettore degli alpha, necessario per la funzione `define_p_treat`}}
}
\description{
Per trovare l'intercetta adeguata al modello di stima dei propensity score per ottenere una determinata treatment prevalence desiderata.
}
