% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix_statin.R
\name{corr_matrix_statin}
\alias{corr_matrix_statin}
\title{{corr_matrix_statin: correlation matrix extraction from case study}}
\usage{
corr_matrix_statin(data)
}
\arguments{
\item{{data}}{{Case study dataset}}
}
\description{
{corr_matrix_statin: correlation matrix extraction from case study}
}
