% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_alpha.R
\name{define_alpha}
\alias{define_alpha}
\title{{Define alpha: define the logistic regression coefficients}}
\usage{
define_alpha(data)
}
\arguments{
\item{{data}}{{Case study dataset}}
}
\description{
{Define alpha: define the logistic regression coefficients}
}
