% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_beta.R
\name{define_beta}
\alias{define_beta}
\title{{Define beta: define the coefficients for simulating the outcome}}
\usage{
define_beta(data)
}
\arguments{
\item{{data}}{{Case study dataset}}
}
\description{
{Define beta: define the coefficients for simulating the outcome}
}
