% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_p_treat.R
\name{define_p_treat}
\alias{define_p_treat}
\title{{Definizione propensity score osservazioni simulate}}
\usage{
define_p_treat(int, alpha, baseline_covariates)
}
\arguments{
\item{{int}}{{Intercetta del modello di regressione logistica: ottenibile utilizzando la funzione `bisection`}}

\item{{alpha}}{{Vettore dei coefficienti di regressione logistica tra le variabili confondenti e la probabilità del trattamento, deve avere la stessa lunghezza del numero di colonne della matrice `baseline_covariates`, ovvero 9}}

\item{{baseline_covariates}}{{Variabili confondenti simulate utilizzando la funzione `baseline_covariates`}}
}
\description{
Definisce la probabilità di ricevere il trattamento per ciascuna delle osservazioni simulate
}
