% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{{Simulation: Monte Carlo simulation}}
\usage{
simulation(
  n,
  n_datasets = 1000,
  seed = 42,
  treatment_prevalence,
  treatment_effect = -0.2,
  approach = c("crude", "IPW", "OW"),
  trim = NULL
)
}
\arguments{
\item{{n}}{{Number of simulated observations}}

\item{{n_datasets}}{{Number of simulated dataset, default 1000}}

\item{{seed}}{{Random seed, default 42}}

\item{{treatment_prevalence}}{{Desired treatment prevalence}}

\item{{treatment_effect}}{{Fixed treatment effect, default -0.9}}

\item{{approach}}{{Type of approach, i.e. crude, IPW, OW}}

\item{{trim}}{{If IPW, for specifying the desired trimming cutoff (1 means no trimming)}}
}
\value{
{A list containing: balancing levels, bias and RMSE}
}
\description{
Simulated a specified configuration
}
