% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_results.R
\name{simulation_results}
\alias{simulation_results}
\title{{Simulation results}}
\usage{
simulation_results(sim_results, treatment_effect = -0.9)
}
\arguments{
\item{{sim_results}}{{Output of `simulation`}}

\item{{treatment_effect}}{{Desired treatment effect, default -0.9}}
}
\description{
Extract the simulation results in terms of Absolute Standardized Mean Difference, Bias and RMSE
}
